test(a11y): axe-core gate for SPA (WCAG 2.1 AA)
Three states scanned: initial render, after a bruto change (chart
loaded, sr-only table populated), and with the Excel button in its
aria-busy="true" state. Asserts zero serious/critical violations
against wcag2a/wcag2aa/wcag21a/wcag21aa/best-practice tags. Pins
data-theme=light so contrast rules are deterministic.

The aria-busy scan sets the attribute directly on the button rather
than clicking it: a real click would trigger the full 15-year x 100k-
row workbook build inside the browser (minutes of CPU). The point of
the scan is to verify the busy-state markup is accessible, not to
exercise SheetJS.

To make this testable without forcing main.ts to re-execute its
top-level mount per test, the bootstrap logic moves to a new
src/ui/app.ts exporting mountApp(target). main.ts shrinks to a thin
shim that finds #app and calls mountApp; the test does the same with
its own DOM. No behaviour change in production.

Refs #52
